consolidates password-reset and corrects uidb64 field typo
