+++ lib/account/password-reset.ts
@@ -1,10 +1,40 @@
 import { jsonCall } from "$http";
+import type { PasswordResetCredentials } from "./types.ts";
 
+/**
+ * Register a new user.
+ *
+ * @returns The raw GET "/account/register" endpoint response
+ */
+export async function passwordResetRequest(
+  phone: string,
+): Promise<PasswordResetCredentials> {
+  return await jsonCall<PasswordResetCredentials>({
+    url: "account/password-reset",
+    init: {
+      method: "POST",
+      headers: {
+        "Accept": "application/json",
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({ phone }),
+    },
+  });
+}
+
+/**
+ * The password reset request configuration.
+ */
 export interface PasswordResetConfig {
+  /** The code send via SMS */
   code: string;
-  uid64: string;
+  /** The base 64 encoded user id recieved by the password reset request */
+  uidb64: string;
+  /** The token string recieved by the password reset request */
   token: string;
+  /** The desired new password */
   new_password1: string;
+  /** The desired new password */
   new_password2: string;
 }
 
@@ -17,9 +47,9 @@ export async function resetPassword(
   config: PasswordResetConfig,
 ): Promise<unknown> {
   //TODO: get return type
-  const { token, uid64, ...form } = config;
+  const { token, uidb64, ...form } = config;
   return await jsonCall<unknown>({
-    url: `account/password-reset/${uid64}/${token}`,
+    url: `account/password-reset/${uidb64}/${token}`,
     init: {
       method: "POST",
       headers: {
