+++ lib/account/schema.ts
@@ -31,13 +31,13 @@ export const userDetailsSchema: UserDetailsSchema = z.object({
 export interface PasswordResetCredentialsSchema extends
   z.ZodObject<{
     /** The ID of the user */
-    uid64: z.ZodString;
+    uidb64: z.ZodString;
     /** The full name of the user */
     token: z.ZodString;
   }> {}
 
 export const passwordResetCredentialsSchema: PasswordResetCredentialsSchema = z
   .object({
-    uid64: z.string(),
+    uidb64: z.string(),
     token: z.string(),
   });
